Wrap coordinates around dateline
When browsing another copy of the world, left or right the central main one, after right-clicking on the map to add start, end, or intermediate points, error is displayed:
Cannot find point 0: 52.50452,373.351135
Cannot find point 1: 52.516639,373.428383

This PR will make all clicks with longitude -180 to +180 degrees.

Uses the [wrap()](http://leafletjs.com/reference.html#latlng-wrap) function from Leaflet.

It's not a complete solution to this issue, as the route is always drawn on the main copy of the world. However, the error is gone, and and the map is automatically panned to the route anyway (in main copy of world).