Guicify
GraphHopperGtfs doesn't inherit from GraphHopper anymore, so it couldn't be used with the Servlets anymore.
Changed the Servlets so they are only injected with what they need.
GraphHopperGtfs is now constructed with pure dependency injection - construct everything bottom-up
and pass it in the constructor.
DefaultModule is retrofitted so it gets the injected things from GraphHopper.