Fixes setting of maxVisitedNodes for witness path searches in PrepareContractionHierarchies.
* so far maxVisitedNodes was always set to zero in prepareNodes() / calculatePriority()
because meanDegree was initialized only afterwards in contractNodes()
this means so far the priorities were not calculated right because no witness
paths could be found.

* this fix requires adjusting three tests that depend on the contraction
order. adjusted createExampleGraph instead of testAddShortcut, because
testAddShortcut does no longer find a shortcut with the example graph.
the other tests using createExampleGraph are not be affected by this change.