Add snap prevention option: pedestrian_area
In order to avoid snapping to the "wrong side" of a long pedestrian area.
The typical pedestrian area (`highway=pedestrian, area=true`) that causes
problems is a long coastal esplanade: There are several connecting edges
from the land side, but none from the water side. If we snap to the water
side of the esplanade, the route will trackaround  the edges of the area
before getting to the connecting land side edges. This results in routes
that are way too long (a route that should be a few meters can become
several kilometers).

One example of such esplanade is Norra Hammarbyhamnen, Stockholm, Sweden.