Make OSMInputFile call callback methods for each read OSM object
When instantiating OSMInputFile the caller has to register at least one
instace of a class implementing the OSMHandler interface. The interface
defines one method for each OSM object type and is inspired by the class
osmium::handler::Handler of the Libosmium C++ library. This commit
moves the switch-case blocks from the class building the graph to a
class which is really related to OSM raw data. It builds the base to
allow adding extensions with additional handlers by downstream
developers without inheritation in future (not possible now but useful to make OSMReader
extensible without inheritation).