fix(pom,cve): Upgrade dropwizard from 2.0.21 to 4.0.1
This introduces a lot of up to date packages but also drops junit < 5 support. Core still relies on that so it needs a separate package import. We should not touch that too much until a newer GH merge.
This fixes mostly the guava CVE CVE-2023-2976 and GHSA-7g45-4rm6-3mm3