fix: support Barrier Nodes for NavigateResponse
Supplying a route with Intersections located on the same spot, connected with a zero length edge seems
to confuse Mapbox like APIs. This can happen if there are barriers on the route, which are split into
two intersections connected by a zero length edge.
The idea is to find these intersections and merge all the edges of both intersections and remove the
out outgoing edge of the connecting edge from the first and the in edge of the incoming edge.