feat: Treat separate sidewalks like foot=use_sidepath
Separately mapping sidewalks becomes more and more common.
In many jurisdictions, available sidewalks must be used, so route
via the separate sidewalks by default.

Several tags are in use related to sidewalks (sidewalk:both, sidewalk:left,
sidewalk:right) with yes/no/separate as options.
If only one side is specified (like sidewalk:left without sidewalk:right),
default behaviour is maintained.
Best practice is to define both or sidewalk:both.
The tag sidewalk=* can additionally define left/right/both indicating
which side has or hasn't a sidewalk and that it is part of the way.

If either side is separate and the other not part of the way, don't route
via the way unless the alternatives are impractical.